+++ chat-client/src/client/mynahUi.ts
@@ -835,6 +835,7 @@ export const createMynahUi = (
         },
         config: {
             maxTabs: 10,
+            test: true,
             dragOverlayIcon: MynahIcons.IMAGE,
             texts: {
                 ...uiComponentsTexts,